+++ relay-server/src/service.rs
@@ -142,6 +142,7 @@ impl ServiceState {
         let project_cache = ProjectCache::new(
             config.clone(),
             event_manager.clone(),
+            outcome_producer.clone(),
             upstream_relay.clone(),
             redis_pool,
         )