+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+**Internal**:
+
+- Emit outcomes for rate limited attachments. ([#951](https://github.com/getsentry/relay/pull/951))
+
 ## 21.3.1
 
 **Bug Fixes**: